SI-6811 Misc. removals in util, testing, io, ...
